Improved testing scripting, added script to run some tests with partial update.

+++ tests/run_some_jobs_scripts_with_partial_update.bash
@@ -2,24 +2,26 @@
 
 cd $(dirname "$0")
 
-rm -rf ./voronota_package
-
-../package.bash voronota_package
-mv ../voronota_package.tar.gz ./voronota_package.tar.gz
-tar -xf ./voronota_package.tar.gz
-rm ./voronota_package.tar.gz
+if [ ! -d "./voronota_package" ]
+then
+	../package.bash voronota_package
+	mv ../voronota_package.tar.gz ./voronota_package.tar.gz
+	tar -xf ./voronota_package.tar.gz
+	rm ./voronota_package.tar.gz
+else
+	cp ../voronota ../voronota-* ./voronota_package/
+fi
 
 export VORONOTA=./voronota_package/voronota
 export VORONOTADIR=./voronota_package/
 export INPUTDIR=./input
 export OUTPUTDIR=./jobs_output
 
-mkdir $VORONOTADIR/resources
 $VORONOTADIR/voronota-resources energy_means_and_sds > $VORONOTADIR/resources/energy_means_and_sds
 $VORONOTADIR/voronota-resources energy_potential > $VORONOTADIR/resources/energy_potential
 $VORONOTADIR/voronota-resources radii > $VORONOTADIR/resources/radii
 
-rm -r -f $OUTPUTDIR
-mkdir $OUTPUTDIR
-
-find ./jobs_scripts/ -type f -name "*.bash" | xargs -L 1 -P 4 bash
+for JOBSCRIPT in "$@"
+do
+	bash "./jobs_scripts/do_${JOBSCRIPT}.bash"
+done
